requirements.txt
beautifulsoup4
certifi
charset-normalizer
colorama
coloredlogs
colorlog
ddddocr
flatbuffers
humanfriendly
idna
mpmath
numpy
onnxruntime
opencv-python-headless
packaging
pillow
protobuf
pyreadline3
python-dotenv
python-rapidjson
pytz
requests
soupsieve
sympy
typing_extensions
ujson
urllib3
subprocess
time
datetime
json
os